requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-crontab==0.7.1
psycopg2-binary==2.8.5
python-decouple
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.2.0